further refactoring, and some debugging
