+++ panther/include/panther_ros.hpp
@@ -24,10 +24,12 @@
 #include <panther_msgs/DynTraj.h>
 
 #include "utils.hpp"
+#include "ros_utils.hpp"
 #include "panther.hpp"
 #include "panther_types.hpp"
 
 #include "timer.hpp"
+#include "ros_timer.hpp"
 
 // #define WHOLE 1  // Whole trajectory (part of which is planned on unkonwn space)
 // #define SAFE 2   // Safe path
