+++ panther/src/ros_utils.cpp
@@ -275,175 +275,6 @@ mt::PieceWisePol createPwpFromStaticPosition(const mt::state& current_state)
   return pwp;
 }
 
-// returns a mt::PieceWisePol, taking the polynomials of p1 and p2 that should satisfy p1>t, p2>t
-mt::PieceWisePol composePieceWisePol(const double t, const double dc, mt::PieceWisePol& p1, mt::PieceWisePol& p2)
-{
-  // if t is in between p1 and p2, force p2[0] to be t
-  if (t > p1.times.back() && t < p2.times.front())  // && fabs(t - p2.times.front()) <= dc) TODO Sometimes fabs(t -
-                                                    // p2.times.front()) is 0.18>c
-  {
-    p2.times.front() = t;
-  }
-
-  if (p1.times.back() < p2.times.front())  // && fabs(p1.times.back() - p2.times.front()) <= dc) TODO
-  {
-    p2.times.front() = p1.times.back();
-  }
-
-  if (t < p1.times.front())  // TODO
-  {
-    p1.times.front() = t;
-  }
-
-  if (fabs(t - p2.times.front()) < 1e-5)
-  {
-    return p2;
-  }
-
-  if (p1.times.back() < p2.times.front() || t > p2.times.back() || t < p1.times.front())
-  {
-    // TODO?
-    // std::cout << "Error composing the piecewisePol" << std::endl;
-    // std::cout << std::setprecision(30) << "t= " << t << std::endl;
-    // std::cout << std::setprecision(30) << "p1.times.front()= " << p1.times.front() << std::endl;
-    // std::cout << std::setprecision(30) << "p1.times.back()= " << p1.times.back() << std::endl;
-    // std::cout << std::setprecision(30) << "p2.times.front() = " << p2.times.front() << std::endl;
-    // std::cout << std::setprecision(30) << "p2.times.back() = " << p2.times.back() << std::endl;
-    mt::PieceWisePol tmp;
-    return tmp;
-  }
-
-  std::vector<int> indexes1, indexes2;
-
-  for (int i = 0; i < p1.times.size(); i++)
-  {
-    if (p1.times[i] > t && p1.times[i] < p2.times[0])
-    {
-      indexes1.push_back(i);
-    }
-  }
-
-  for (int i = 0; i < p2.times.size(); i++)
-  {
-    if (p2.times[i] > t)
-    {
-      indexes2.push_back(i);
-    }
-  }
-
-  mt::PieceWisePol p;
-  p.times.push_back(t);
-
-  for (auto index_1_i : indexes1)
-  {
-    p.times.push_back(p1.times[index_1_i]);
-    p.all_coeff_x.push_back(p1.all_coeff_x[index_1_i - 1]);
-    p.all_coeff_y.push_back(p1.all_coeff_y[index_1_i - 1]);
-    p.all_coeff_z.push_back(p1.all_coeff_z[index_1_i - 1]);
-  }
-
-  for (auto index_2_i : indexes2)
-  {
-    if (index_2_i == 0)
-    {
-      p.all_coeff_x.push_back(p1.all_coeff_x.back());
-      p.all_coeff_y.push_back(p1.all_coeff_y.back());
-      p.all_coeff_z.push_back(p1.all_coeff_z.back());
-      p.times.push_back(p2.times[index_2_i]);
-      continue;
-    }
-    p.times.push_back(p2.times[index_2_i]);
-    p.all_coeff_x.push_back(p2.all_coeff_x[index_2_i - 1]);
-    p.all_coeff_y.push_back(p2.all_coeff_y[index_2_i - 1]);
-    p.all_coeff_z.push_back(p2.all_coeff_z[index_2_i - 1]);
-  }
-
-  return p;
-}
-
-std::vector<std::string> pieceWisePol2String(const mt::PieceWisePol& pwp)
-{
-  // Define strings
-  std::string s_x = "0.0";
-  std::string s_y = "0.0";
-  std::string s_z = "0.0";
-
-  int deg = pwp.getDeg();
-
-  //(pwp.times - 1) is the number of intervals
-  for (int i = 0; i < (pwp.times.size() - 1); i++)  // i is the index of the interval
-  {
-    std::string div_by_delta = "/ (" + std::to_string(pwp.times[i + 1] - pwp.times[i]) + ")";
-
-    std::string t = "(min(t," + std::to_string(pwp.times.back()) + "))";
-
-    std::string u = "(" + t + "-" + std::to_string(pwp.times[i]) + ")" + div_by_delta;
-    u = "(" + u + ")";
-
-    // std::string u = "(" + t + "-" + std::to_string(pwp.times[i]) + ")" + div_by_delta;
-    // u = "(" + u + ")";
-    // std::string uu = u + "*" + u;
-    // std::string uuu = u + "*" + u + "*" + u;
-    /*    std::cout << "pwp.times[i]= " << pwp.times[i] << std::endl;
-        std::cout << "pwp.times[i+1]= " << pwp.times[i + 1] << std::endl;*/
-
-    std::string cond;
-    if (i == (pwp.times.size() - 2))  // if the last interval
-    {
-      cond = "(t>=" + std::to_string(pwp.times[i]) + ")";
-    }
-    else
-    {
-      cond = "(t>=" + std::to_string(pwp.times[i]) + " and " + "t<" + std::to_string(pwp.times[i + 1]) + ")";
-    }
-
-    std::string s_x_i = "";
-    std::string s_y_i = "";
-    std::string s_z_i = "";
-    for (int j = 0; j <= deg; j++)
-    {
-      std::string power_u = "(" + u + "^" + std::to_string(deg - j) + ")";
-
-      s_x_i = s_x_i + "+" + std::to_string((double)pwp.all_coeff_x[i](j)) + "*" + power_u;
-      s_y_i = s_y_i + "+" + std::to_string((double)pwp.all_coeff_y[i](j)) + "*" + power_u;
-      s_z_i = s_z_i + "+" + std::to_string((double)pwp.all_coeff_z[i](j)) + "*" + power_u;
-    }
-
-    s_x_i = cond + "*(" + s_x_i + ")";
-    s_y_i = cond + "*(" + s_y_i + ")";
-    s_z_i = cond + "*(" + s_z_i + ")";
-
-    // std::string s_x_i = std::to_string((double)pwp.all_coeff_x[i](0)) + "*" + uuu;   //////////////////
-    // s_x_i = s_x_i + "+" + std::to_string((double)pwp.all_coeff_x[i](1)) + "*" + uu;  //////////////////
-    // s_x_i = s_x_i + "+" + std::to_string((double)pwp.all_coeff_x[i](2)) + "*" + u;   //////////////////
-    // s_x_i = s_x_i + "+" + std::to_string((double)pwp.all_coeff_x[i](3));             //////////////////
-    // s_x_i = cond + "*(" + s_x_i + ")";
-
-    // std::string s_y_i = std::to_string((double)pwp.all_coeff_y[i](0)) + "*" + uuu;   //////////////////
-    // s_y_i = s_y_i + "+" + std::to_string((double)pwp.all_coeff_y[i](1)) + "*" + uu;  //////////////////
-    // s_y_i = s_y_i + "+" + std::to_string((double)pwp.all_coeff_y[i](2)) + "*" + u;   //////////////////
-    // s_y_i = s_y_i + "+" + std::to_string((double)pwp.all_coeff_y[i](3));             //////////////////
-    // s_y_i = cond + "*(" + s_y_i + ")";
-
-    // std::string s_z_i = std::to_string((double)pwp.all_coeff_z[i](0)) + "*" + uuu;   //////////////////
-    // s_z_i = s_z_i + "+" + std::to_string((double)pwp.all_coeff_z[i](1)) + "*" + uu;  //////////////////
-    // s_z_i = s_z_i + "+" + std::to_string((double)pwp.all_coeff_z[i](2)) + "*" + u;   //////////////////
-    // s_z_i = s_z_i + "+" + std::to_string((double)pwp.all_coeff_z[i](3));             //////////////////
-    // s_z_i = cond + "*(" + s_z_i + ")";
-
-    s_x = s_x + " + " + s_x_i;
-    s_y = s_y + " + " + s_y_i;
-    s_z = s_z + " + " + s_z_i;
-  }
-
-  std::vector<std::string> s;
-  s.push_back(s_x);
-  s.push_back(s_y);
-  s.push_back(s_z);
-
-  return s;
-}
-
 geometry_msgs::Pose identityGeometryMsgsPose()
 {
   geometry_msgs::Pose pose;
