+++ panther_compression/compression/utils/ActionManager.py
@@ -1,7 +1,13 @@
 import numpy as np
+import numpy.matlib
 import py_panther
 
 from colorama import Fore, Style
+
+from .MyClampedUniformBSpline import MyClampedUniformBSpline
+from .utils import generateKnotsForClampedUniformBspline, numpy3XmatrixToListOf3dVectors
+from .State import State
+
 from .yaml_utils import readPANTHERparams
 
 class ActionManager():
