+++ panther_compression/compression/utils/ObservationManager.py
@@ -2,8 +2,11 @@ import numpy as np
 import py_panther
 
 from .yaml_utils import readPANTHERparams
-
+from .utils import normalize, wrapInmPiPi
+from .State import State
 from .ObstaclesManager import ObstaclesManager
+
+from colorama import Fore, Style
 
 class ObservationManager():
 	def __init__(self):
