+++ panther_compression/compression/utils/Obstacle.py
@@ -1,0 +1,4 @@
+class Obstacle():
+	def __init__(self, ctrl_pts, bbox_inflated):
+		self.ctrl_pts=ctrl_pts
+		self.bbox_inflated=bbox_inflated
